refactor(core): ISS-008/009/010 extract shared parsing_utils.hpp, remove ~200 LOC duplication
